イベント機能の基本部分を実装しました。

+++ chimera-app/app/components/memo/memo-delete-confirm-dialog.tsx
@@ -1,49 +1,68 @@
 import * as React from 'react'
 import { Form } from '@remix-run/react'
 import { useTranslation } from 'react-i18next'
-import { AlertDialogCancel } from '~/components/ui/alert-dialog'
-import { Button } from '~/components/ui/button'
-import { DeleteConfirmDialog } from '~/components/lib/delete-confirm-dialog'
+import {
+  AlertDialogCancel,
+  AlertDialogAction,
+} from '~/components/ui/alert-dialog'
+import { buttonVariants } from '~/components/ui/button'
+import { MEMO_URL } from '~/constants'
+import { ConfirmDialog } from '~/components/lib/confirm-dialog'
 import { Memo } from '~/types/memos'
 
-interface DeleteMemoConfirmDialogProps {
-  memo: Memo
-  isOpen: boolean
-  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>
+export interface DeleteMemoConfirmDialogProps {
+  memo: Memo | undefined
+  isOpen?: boolean
+  setIsOpen?: React.Dispatch<React.SetStateAction<boolean>>
+  onSubmit?: (event: React.FormEvent<HTMLFormElement>) => void
+  returnUrl?: string
+  children?: React.ReactNode
 }
 
 export function MemoDeleteConfirmDialog({
   memo,
   isOpen,
   setIsOpen,
+  onSubmit,
+  returnUrl = MEMO_URL,
+  children,
 }: DeleteMemoConfirmDialogProps) {
   const { t } = useTranslation()
 
+  if (!memo) return null
+
+  const action = [MEMO_URL, memo.id, 'delete'].join('/')
+  const desc =
+    '「' +
+    (memo.title || t('memo.un_titled')) +
+    '」' +
+    t('common.message.confirm_deletion')
+
   return (
-    <DeleteConfirmDialog
+    <ConfirmDialog
       title={t('memo.message.memo_deletion')}
-      description={
-        '「' +
-        (memo.title || t('memo.un_titled')) +
-        '」' +
-        t('common.message.confirm_deletion')
-      }
+      description={desc}
       isOpen={isOpen}
       setIsOpen={setIsOpen}
+      torigger={children}
     >
       <AlertDialogCancel>{t('common.message.cancel')}</AlertDialogCancel>
-      <Form
-        action={`/memos/${memo.id}/delete`}
-        method="delete"
-        onSubmit={(event) => {
-          setIsOpen(false)
-          event.stopPropagation()
-        }}
+      {/* レスポンシブ対応のためにFormでButtonを囲まない */}
+      <AlertDialogAction
+        type="submit"
+        className={buttonVariants({ variant: 'destructive' })}
+        form="delete-memo-form"
       >
-        <Button type="submit" variant="destructive">
-          {t('common.message.delete')}
-        </Button>
+        {t('common.message.delete')}
+      </AlertDialogAction>
+      <Form
+        id="delete-memo-form"
+        action={action}
+        method="delete"
+        onSubmit={onSubmit}
+      >
+        <input type="hidden" name="returnUrl" value={returnUrl} />
       </Form>
-    </DeleteConfirmDialog>
+    </ConfirmDialog>
   )
 }
